requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.15
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.4
textblob==0.17.1
numpy==1.24.3
transformers==4.31.0
torch==2.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
psutil==5.9.5
google-generativeai==0.3.0
nltk==3.8.1
fastapi==0.104.1
uvicorn==0.23.2
openai==1.5.0
aioredis==2.0.1
fastapi-limiter==0.1.5
pydantic==2.4.2
httpx==0.25.0
cryptography==44.0.3